% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{cisnienie_hPa}
\alias{cisnienie_hPa}
\title{Konwersja cisnienia (hektopaskale)}
\usage{
cisnienie_hPa(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc hektopaskali}
}
\value{
data frame z pozostalymi jednostkami cisnienia i ich wartosci
}
\description{
Funkcja sluzaca do konwersji hektopaskali na pozostale jednostki cisnienia
}
\examples{
cisnienie_hPa(20)
cisnienie_hPa(c(40, 95))
}
