% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{cisnienie_kgcm2}
\alias{cisnienie_kgcm2}
\title{Konwersja cisnienia (kilogramy na centymetr kwadratowy)}
\usage{
cisnienie_kgcm2(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc kilogramow na centymetr kwadratowy}
}
\value{
data frame z pozostalymi jednostkami cisnienia i ich wartosci
}
\description{
Funkcja sluzaca do konwersji kilogramow na centymetr kwadratowy na pozostale jednostki cisnienia
}
\examples{
cisnienie_kgcm2(20)
cisnienie_kgcm2(c(40, 95))
}
