% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{cisnienie_mmWg}
\alias{cisnienie_mmWg}
\title{Konwersja cisnienia (milimetr slupa wody)}
\usage{
cisnienie_mmWg(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc milimetrow slupa wody}
}
\value{
data frame z pozostalymi jednostkami cisnienia i ich wartosci
}
\description{
Funkcja sluzaca do konwersji milimetrow slupa wody na pozostale jednostki cisnienia
}
\examples{
cisnienie_mmWg(20)
cisnienie_mmWg(c(40, 95))
}
