% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{cisnienie_Pa}
\alias{cisnienie_Pa}
\title{Konwersja cisnienia (paskal)}
\usage{
cisnienie_Pa(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc paskali}
}
\value{
data frame z pozostalymi jednostkami cisnienia i ich wartosci
}
\description{
Funkcja sluzaca do konwersji paskali na pozostale jednostki cisnienia
}
\examples{
cisnienie_Pa(20)
cisnienie_Pa(c(40, 95))
}
