% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{dlugosc_dm}
\alias{dlugosc_dm}
\title{Konwersja dlugosci (decymetry)}
\usage{
dlugosc_dm(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc decymetrow}
}
\value{
data frame z pozostalymi jednostkami dlugosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji decymetrow na pozostale jednostki dlugosci
}
\examples{
dlugosc_dm(20)
dlugosc_dm(c(40, 95))
}
