% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{dlugosc_ft}
\alias{dlugosc_ft}
\title{Konwersja dlugosci (stopy)}
\usage{
dlugosc_ft(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc stop}
}
\value{
data frame z pozostalymi jednostkami dlugosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji stop na pozostale jednostki dlugosci
}
\examples{
dlugosc_ft(20)
dlugosc_ft(c(40, 95))
}
