% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{dlugosc_mm}
\alias{dlugosc_mm}
\title{Konwersja dlugosci (milimetry)}
\usage{
dlugosc_mm(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc milimetrow}
}
\value{
data frame z pozostalymi jednostkami dlugosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji milimetrow na pozostale jednostki dlugosci
}
\examples{
dlugosc_mm(20)
dlugosc_mm(c(40, 95))
}
