% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{dlugosc_nmi}
\alias{dlugosc_nmi}
\title{Konwersja dlugosci (mila morska)}
\usage{
dlugosc_nmi(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc mil morskich}
}
\value{
data frame z pozostalymi jednostkami dlugosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji mil morskich na pozostale jednostki dlugosci
}
\examples{
dlugosc_nmi(20)
dlugosc_nmi(c(40, 95))
}
