% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{dlugosc_yd}
\alias{dlugosc_yd}
\title{Konwersja dlugosci (jardy)}
\usage{
dlugosc_yd(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc jardow}
}
\value{
data frame z pozostalymi jednostkami dlugosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji jardow na pozostale jednostki dlugosci
}
\examples{
dlugosc_yd(20)
dlugosc_yd(c(40, 95))
}
