% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{masa_g}
\alias{masa_g}
\title{Konwersja masy (gramy)}
\usage{
masa_g(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc gramow}
}
\value{
data frame z pozostalymi jednostkami masy i ich wartosci
}
\description{
Funkcja sluzaca do konwersji gramow na pozostale jednostki masy
}
\examples{
masa_g(20)
masa_g(c(40, 95))
}
