% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{masa_lb}
\alias{masa_lb}
\title{Konwersja masy (funty)}
\usage{
masa_lb(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc funtow}
}
\value{
data frame z pozostalymi jednostkami masy i ich wartosci
}
\description{
Funkcja sluzaca do konwersji funtow na pozostale jednostki masy
}
\examples{
masa_lb(20)
masa_lb(c(40, 95))
}
