% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{masa_mg}
\alias{masa_mg}
\title{Konwersja masy (miligramy)}
\usage{
masa_mg(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc miligramow}
}
\value{
data frame z pozostalymi jednostkami masy i ich wartosci
}
\description{
Funkcja sluzaca do konwersji miligramow na pozostale jednostki masy
}
\examples{
masa_mg(20)
masa_mg(c(40, 95))
}
