% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{masa_oz}
\alias{masa_oz}
\title{Konwersja masy (uncje)}
\usage{
masa_oz(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc uncji}
}
\value{
data frame z pozostalymi jednostkami masy i ich wartosci
}
\description{
Funkcja sluzaca do konwersji uncji na pozostale jednostki masy
}
\examples{
masa_oz(20)
masa_oz(c(40, 95))
}
