% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{masa_ton_us}
\alias{masa_ton_us}
\title{Konwersja masy (tona amerykanska)}
\usage{
masa_ton_us(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc ton amerykanskich}
}
\value{
data frame z pozostalymi jednostkami masy i ich wartosci
}
\description{
Funkcja sluzaca do konwersji ton amerykanskich na pozostale jednostki masy
}
\examples{
masa_ton_us(20)
masa_ton_us(c(40, 95))
}
