% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_barylka}
\alias{objetosc_barylka}
\title{Konwersja objetosci (barylki)}
\usage{
objetosc_barylka(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc barylek}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji barylek na pozostale jednostki objetosci
}
\examples{
objetosc_barylka(20)
objetosc_barylka(c(40, 95))
}
