% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_cm3}
\alias{objetosc_cm3}
\title{Konwersja objetosci (centymetry szescienne)}
\usage{
objetosc_cm3(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc centymetrow szesciennych}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji centymetrow szesciennych na pozostale jednostki objetosci
}
\examples{
objetosc_cm3(20)
objetosc_cm3(c(40, 95))
}
