% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_ft3}
\alias{objetosc_ft3}
\title{Konwersja objetosci (stopy szescienne)}
\usage{
objetosc_ft3(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc stop szesciennych}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji stop szesciennych na pozostale jednostki objetosci
}
\examples{
objetosc_ft3(20)
objetosc_ft3(c(40, 95))
}
