% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_galon_uk}
\alias{objetosc_galon_uk}
\title{Konwersja objetosci (galon brytyjski)}
\usage{
objetosc_galon_uk(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc galonow brytyjskich}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji galonow brytyjskich na pozostale jednostki objetosci
}
\examples{
objetosc_galon_uk(20)
objetosc_galon_uk(c(40, 95))
}
