% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_galon_us}
\alias{objetosc_galon_us}
\title{Konwersja objetosci (galon plynny amerykanski)}
\usage{
objetosc_galon_us(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc galonow plynnych, amerykanskich}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji galonu plynnego, amerykanskiego na pozostale jednostki objetosci
}
\examples{
objetosc_galon_us(20)
objetosc_galon_us(c(40, 95))
}
