% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_kwarta_uk}
\alias{objetosc_kwarta_uk}
\title{Konwersja objetosci (kwarta brytyjska)}
\usage{
objetosc_kwarta_uk(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc kwarty brytyjskiej}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji kwarty brytyjskiej na pozostale jednostki objetosci
}
\examples{
objetosc_kwarta_uk(20)
objetosc_kwarta_uk(c(40, 95))
}
