% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_kwarta_us}
\alias{objetosc_kwarta_us}
\title{Konwersja objetosci (kwarta amerykanska)}
\usage{
objetosc_kwarta_us(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc kwarty amerykanskiej}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji kwarty amerykanskiej na pozostale jednostki objetosci
}
\examples{
objetosc_kwarta_us(20)
objetosc_kwarta_us(c(40, 95))
}
