% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_m3}
\alias{objetosc_m3}
\title{Konwersja objetosci (metry szescienne)}
\usage{
objetosc_m3(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc metrow szesciennych}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji metrow szesciennych na pozostale jednostki objetosci
}
\examples{
objetosc_m3(20)
objetosc_m3(c(40, 95))
}
