% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_uncja_us}
\alias{objetosc_uncja_us}
\title{Konwersja objetosci (uncje amerykanskie)}
\usage{
objetosc_uncja_us(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc uncji amerykanskich}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji uncji amerykanskich na pozostale jednostki objetosci
}
\examples{
objetosc_uncja_us(20)
objetosc_uncja_us(c(40, 95))
}
