% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{objetosc_yd3}
\alias{objetosc_yd3}
\title{Konwersja objetosci (jardy szescienne)}
\usage{
objetosc_yd3(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc jardow szesciennych}
}
\value{
data frame z pozostalymi jednostkami objetosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji jardow szesciennych na pozostale jednostki objetosci
}
\examples{
objetosc_yd3(20)
objetosc_yd3(c(40, 95))
}
