% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{powierzchnia_a}
\alias{powierzchnia_a}
\title{Konwersja powierzchni (ary)}
\usage{
powierzchnia_a(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc arow}
}
\value{
data frame z pozostalymi jednostkami powierzchni i ich wartosci
}
\description{
Funkcja sluzaca do konwersji arow na pozostale jednostki powierzchni
}
\examples{
powierzchnia_a(20)
powierzchnia_a(c(40, 95))
}
