% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{powierzchnia_acre}
\alias{powierzchnia_acre}
\title{Konwersja powierzchni (akry)}
\usage{
powierzchnia_acre(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc akrow}
}
\value{
data frame z pozostalymi jednostkami powierzchni i ich wartosci
}
\description{
Funkcja sluzaca do konwersji akrow na pozostale jednostki powierzchni
}
\examples{
powierzchnia_acre(20)
powierzchnia_acre(c(40, 95))
}
