% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{powierzchnia_mm2}
\alias{powierzchnia_mm2}
\title{Konwersja powierzchni (milimetry kwadratowe)}
\usage{
powierzchnia_mm2(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc milimetrow kwadratowych}
}
\value{
data frame z pozostalymi jednostkami powierzchni i ich wartosci
}
\description{
Funkcja sluzaca do konwersji milimetrow kwadratowych na pozostale jednostki powierzchni
}
\examples{
powierzchnia_mm2(20)
powierzchnia_mm2(c(40, 95))
}
