% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{predkosc_kmh}
\alias{predkosc_kmh}
\title{Konwersja predkosci (kilometry na godzine)}
\usage{
predkosc_kmh(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc kilometrow na godzine}
}
\value{
data frame z pozostalymi jednostkami predkosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji kilometrow na godzine na pozostale jednostki predkosci
}
\examples{
predkosc_kmh(20)
predkosc_kmh(c(40, 95))
}
