% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{predkosc_mph}
\alias{predkosc_mph}
\title{Konwersja predkosci (mile na godzine)}
\usage{
predkosc_mph(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc mil na godzine}
}
\value{
data frame z pozostalymi jednostkami predkosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji mil na godzine na pozostale jednostki predkosci
}
\examples{
predkosc_mph(20)
predkosc_mph(c(40, 95))
}
