% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{predkosc_ms}
\alias{predkosc_ms}
\title{Konwersja predkosci (metry na sekunde)}
\usage{
predkosc_ms(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc metrow na sekunde}
}
\value{
data frame z pozostalymi jednostkami predkosci i ich wartosci
}
\description{
Funkcja sluzaca do konwersji metrow na sekunde na pozostale jednostki predkosci
}
\examples{
predkosc_ms(20)
predkosc_ms(c(40, 95))
}
