% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{temperatura_C}
\alias{temperatura_C}
\title{Konwersja temperatury (stopnie Celcjusza)}
\usage{
temperatura_C(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc stopni Celcjusza}
}
\value{
data frame z pozostalymi jednostkami temperatury i ich wartosci
}
\description{
Funkcja sluzaca do konwersji stopni Celcjusza na pozostale jednostki temperatury
}
\examples{
temperatura_C(20)
temperatura_C(c(40, 95))
}
