% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{temperatura_K}
\alias{temperatura_K}
\title{Konwersja temperatury (Kelviny)}
\usage{
temperatura_K(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc Kelvinow}
}
\value{
data frame z pozostalymi jednostkami temperatury i ich wartosci
}
\description{
Funkcja sluzaca do konwersji Kelvinow na pozostale jednostki temperatury
}
\examples{
temperatura_K(20)
temperatura_K(c(40, 95))
}
