% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{temperatura_N}
\alias{temperatura_N}
\title{Konwersja temperatury (stopnie Newtona)}
\usage{
temperatura_N(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc stopni Newtona}
}
\value{
data frame z pozostalymi jednostkami temperatury i ich wartosci
}
\description{
Funkcja sluzaca do konwersji stopni Newtona na pozostale jednostki temperatury
}
\examples{
temperatura_N(20)
temperatura_N(c(40, 95))
}
