% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jednostki_package.R
\name{temperatura_R}
\alias{temperatura_R}
\title{Konwersja temperatury (stopnie Rankinea)}
\usage{
temperatura_R(x)
}
\arguments{
\item{x}{argument zawierajacy wartosc stopni Rankinea}
}
\value{
data frame z pozostalymi jednostkami temperatury i ich wartosci
}
\description{
Funkcja sluzaca do konwersji stopni Rankinea na pozostale jednostki temperatury
}
\examples{
temperatura_R(20)
temperatura_R(c(40, 95))
}
